design : follow 아이템 마크업 (#1)
